Create Shows1-1.html
This is a page for the Nam Entertainment Shows. The page consist of different contents. First, i created the links to other pages, then the picture slideshows which shows contents which are available on the page created using HTML, CSS and JavaScript which are in script. Then i created a group of Shows that the page will be showing in list format and also in pictures. Lastly i created a the page footer that will be showing the contact details and social media icons on which the page is available.